\name{evar-package}
\alias{evar-package}
\docType{package}
\title{
\packageTitle{evar}
}
\description{
\packageDescription{evar}
}
\details{
\tabular{ll}{
Package: \tab evar\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-09-7\cr
License: \tab GPL (>= 2)\cr
}
}
\references{
Tan, X. and Liu, X. (2020). Testing value at risk in ultra-high dimensional expectile models. Manuscript.
}

\keyword{ package }
